requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-filter==21.1
django-cors-headers==3.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2==2.8.6
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
